Add tests for example app
