+++ project/Dependencies.scala
@@ -66,4 +66,7 @@ object Dependencies {
       _ % mUnitVersion
     ) ++ Seq("org.typelevel" %% "discipline-munit" % disciplineMUnitVersion)
 
+  lazy val catsEffectMUnitVersion = "1.0.6"
+  lazy val catsEffectMUnit = Seq("org.typelevel" %% "munit-cats-effect-3" % catsEffectMUnitVersion)
+
 }
